When xfsdump was modified in 2.2.13, the ioctl used to backspace over the stream terminator was changed from MTBSF to MTBSFM to match the ioctl description in the old MT(1L) man page.  The behavior of the MTBSF[M] ioctls have recently changed in Linux as documented in the latest MT(1) man page.
Old Behavior:
   MTBSF  - Backward space count files.  The tape is positioned on the first
            block of the next file.
   MTBSFM - Backward space count file marks.  The tape is positioned on the
            beginning-of-tape side (in front) of the file mark.
New Behavior:
   MTBSF  - Backward space count files.  The tape is positioned on the last
            block of the previous file.
   MTBSFM - Backward space count files.  The tape is positioned on the first
            block of the next file.

The code was changed to MTBSFM in order to end up on the BOT (left) side of
the filemark to overwrite the terminator.  Due to the change in ioctl behavior,
MTBSFM now positions us on the first block of the next file. This means the
stream terminator is never getting overwritten and xfsdump thinks there is only
ever one dump on the tape (and overwrites any dumps previously appended to the
first dump).

The TS tape driver code has been fixed to match the new ioctl behavior, and
this mod will change the xfsdump code back to use MTBSF once again.  However,
APD 2.8 (the fix) won't be released until spring 2004... and TS backups won't
work properly with 2.2.15 without the fix in APD 2.8.  For this reason, this
mod also adds a temporary workaround to call the 'real' TS status ioctl,
MTIOCGET_SGI, rather than the ST-emulated status ioctl, MTIOCGET (with bug).
Once MTIOCGET_SGI has been called to get the correct status of the TS device,
a new routine called map_ts_status() must be used to map the TS mt_dposn
field back to the ST mt_gstat for xfsdump to understand the binary flags.