Add support for dumping and restoring project ids for regular files and directories. This is done in a backwards compatible way -- restores from old IRIX dumps will restore the correct project id, and restores from old Linux dumps will restore a 0 for the project id.
Fix a bug on IRIX where the first two bytes of bs_pad1 in a bstat_t
were not being zeroed before being dumped. Document this fact since
it might affect the ability to use them in the future.

xfsrestore on IRIX was not restoring any extended inode flags for
directories. This is now fixed. Change setdirattr() to do only one
path_to_handle/open_by_handle sequence when restoring directory
attributes. Previously this was being done twice on Linux -- once
for setting the DMAPI event mask and once for the extended inode
flags.