Improve subtree dump performance. The old way of doing subtree dumps was to scan the entire filesystem and add all inodes to the inomap. The non-selected subtrees were then recursed over and those inos were effectively removed from the inomap. The amount of take this takes if related to the size of the filesystem, not the size of the subtree(s) being dumped.
The new way of building the inomap for subtree dumps
is to recurse over the selected subtrees and adding
each ino to the dump. There's no need to do a full
filesystem scan to build the inomap, and no need to
scan files that are not in the selected subtree. As
a result this tends to be a 5-10x speedup over the
old method, depending on the size of the subtree
relative to the size of the filesystem.