The DMF-specific code in xfsdump was taking a shortcut in estimating the dump size of a file in order to avoid the cost of reading the DMF extended attribute. This caused migrating (MIG) files to be identified as dual-state (DUL) files, which lead to an estimated size of 0 for MIG files rather than their true size.
In general it's okay for the estimate to be wrong, but there
are a couple of cases where the estimate needs to be as
accurate as possible, even if it means slowing down the
initial inode scan. These cases are 1) multiple dump streams,
since the estmated size is used to determine where to split
the dump, and 2) when excluding files over a certain size
from the dump. Only the second case applies on Linux, since
multiple dump streams are not currently supported.

This mod changes the DMF-specific code to get an accurate
dump size estimate when either of the above cases are in
effect.