xfsrestore: cache path lookups
In order to resolve a pathname, xfsrestore must work from an inode
number (from the dump) and recurse up the directory entry tree that
it has constructed. Each level of recursion requires a seek and read
to get the name of the dirent, and possibly a mmap of a section of
the directory entry tree if it is not already mapped (and in that
case, possibly a munmap of another section). It's quite common to
resolve pathnames in the same directory consecutively, so simply
caching the parent directory pathname from the previous lookup saves
quite a bit of overhead.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Alex Elder <aelder@sgi.com>