xfsrestore: cleanup node allocation
Simplify the node allocation code. The current code takes some
number of nodes from a new segment and links them into the freelist
whenever the freelist is depleted. There's no reason to put the new
nodes on the freelist, we can just allocate the next available new
node as needed. This also saves a trip through win_map/win_unmap if
there are no nodes on the freelist (the common case).

Newly allocated nodes are no longer zeroed in this code, but all
allocations are done through Node_alloc which does initialize each
field.

Also remove a couple of TREE_DEBUG messages that were not useful.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Acked-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Alex Elder <aelder@sgi.com>