xfsrestore: fix node table setup
The node table setup code unintentionally limits the amount of
virtual memory that will be used for the node table. Rather than
setting a limit based on the remaining virtual memory, the node
table is limited to the amount of memory it thinks it will need
based on the dump's inode count. But the node table stores directory
entries, not inodes, and so dumps with lots of hard links end up
thrashing mapping and unmapping segments of the node table to stay
within the self-imposed virtual memory limit.

This patch also changes the node table to ensure that there are a
power of two nodes per segment. Profiling revealed that while
building the node table, 50% of the restore cpu time was spent doing
division and modulo to convert a node index into its segment index
and offset within the segment. This change prepares the node table
for another patch which will change the lookup mechanism to use
shift and bitwise-and.

Also don't bother passing the estimated segment table size to the
window abstraction. It has to deal with resizing the table anyway
and can choose a reasonable starting size.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Acked-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Alex Elder <aelder@sgi.com>