xfsrestore: check for compatible xfsrestore
When resuming a restore or doing a cumulative restore, xfsrestore
reads state information left around by the previous invocation.
This patch adds logic to determine whether or not restore is able to
make sense of the sense information.

The xfsrestore man page has also been updated to make the user aware
of the requirement to use a compatible restore and system when
resuming restores.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Acked-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Alex Elder <aelder@sgi.com>