xfsdump: rework dialog timeout and EINTR reliance
The current dialog code uses alarm(2) to implement the dialog timeout.
This is suboptimal as the main loop already makes use of alarm.

Additionally the existing code relies on its blocking read(2) being
interrupted if a signal is received. This is not guaranteed to work
as a signal may be delivered to a different thread.

Both of these issues can be resolved by using select(2) rather than
a blocking read.

'dlog_allowed_flag' and 'dlog_signo_received' are now marked 'volatile'
to prevent them from being optimized out from being checked during
a loop.

This also changes dlog_desist() to not change 'dlog_ttyfd', as that
could impact a dialog already in progress. Clearing 'dlog_allowed_flag'
is sufficient to disable dialogs.

Finally add an assert to dlog_string_query() to check that dialogs
are allowed to make it consistent with other promptinput() callers.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Signed-off-by: Alex Elder <aelder@sgi.com>