xfsdump: remove redundant error code mapping
Currently there is both an exit_strings array and an exit_codestring()
function for mapping an exit code to a string. They are very similar
except the latter prefaces every string with "EXIT_" and the former
uses "SUCCESS" instead of "NORMAL", and cannot handle an invalid exit
code. exit_codestring() is not currently used in active code (it's for
multi-stream), so its values can be changed to allow the exit_strings
array to be removed.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Signed-off-by: Alex Elder <aelder@sgi.com>