xfsdump: incremental dump directory selection
When doing incremental backups, xfsdump does a recursive scan through
the filesystems to determine the proper set of unchanged directories
to include in the dump in order to make the dump self-contained
(allowing files to be restored from this dump without applying the
base dumps first).

For high-inode count filesystems, the recursive scan may consume a
significant portion of the backup time. This patch adds a -D option
which causes xfsdump to skip the recursive scan. Unchanged directories
will not be dumped, so the base dump(s) will have to be loaded prior
to restoring files from the dump. When restoring files from such a
dump, a message is issued to alert the user that files may end up in
the orphanage if the base dump(s) has not been applied.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>