xfsdump: zero bs_forkoff, don't fill in the value
In xfsdump 3.1.2 I explicitly added the bs_forkoff member
to this structure; I tried to be good and explicitly fill
in a value.  However, previously it was initialized to
zero, (by virtue of being missing) and now we're giving it
a value (which is ignored by restore, other than to checksum
it).

By putting in a non-zero value, I broke checksumming
when an xfsdump with a non-zero forkoff was restored
by an older xfsrestore that doesn't know about the field.
Fill in 0 to fix backwards compatibility.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reported-by: Fugazzi99 <fugazzi99@gmail.com>
Reviewed-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>