xfsdump: Fix memory and fd leak in invutil/stobj.c
The function open_stobj duplicates its argument, upon successful
duplication, the fstat is called. If the fstat command fails then
the memory for the duplicated string is leaked. Fix this by moving
the string duplication after the fstat call. This is ok because
the fstat call does not use the duplicated string.

Brian Foster noticed that the function also leaks a file descriptor
in case the file cannot be fstated. Fixing that, too.

Signed-off-by: Boris Ranto <ranto.boris@gmail.com>
Reviewed-by: Brian Foster <bfoster@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>