cleanup: get rid of ASSERT
ASSERT comes from the xfs/xfs.h include, and we don't ever define
DEBUG so we never get asserts built in. We want asserts built in for
testing, but not for distro packages. The debian package already
tries to do this by using "export DEBUG=-DNDEBUG" for the build
context, but seeing as we pull in #define ASSERT(ex) (0) from the
XFS headers it's a no-op.

Convert all the ASSERT calls to assert to remove this conflict with
the xfsprogs headers and so local developer builds are built with
asserts enabled. 

Signed-off-by: Dave Chinner <dchinner@redhat.com>