build: don't rely on xfs/xfs.h to include necessary headers
From xfsprogs 4.2.0, the xfs headers no long include lots of
xfsprogs specific build information. xfsdump hasbeen implicitly
relying on those header for a clean build, rather than having it's
own definitions. Hence we need to add the missing bits to
config.h.in and include it in all the files that require the
specific help that the xfs headers used to provide.

Signed-off-by: Dave Chinner <dchinner@redhat.com>